requirements.txt
flask
python-dotenv
nextcord